test(main.go): Begin refactor for testing
This commit starts to refactor the package to use stronger test
techniques.  Two specific changes are implemented in this first
commit toward stronger testing.

Nothing here should break backward compatability.

1. Added lockness_testing.go.  Also deleted the example_testing.go
because it relied on external file dependencies and did not test
the creation of foundatinal types in the package.
2. Added NewLLReaderRequest(config io.Reader) which has a different
function signature than the original NewLLRequest.  The new
function signature returns only the *LLRequest instead of a tuple
that also returned a *JSONDB.  More importantly, the argument to
create a new LLRequest is now an io.Reader instead of a string
filename.  This is critical because it is much easier to create a
buffer with an io.Reader in testing, and then users can pass an
io.Reader created from a file in production.  Because of this
change testing for two error pathways and the success pathway are
now validated in test code.