test(NewLLReaderRequest): Expand test coverage
This commit expands the test coverage for creating a new LLRequest
by covering the case when the config file has yaml tags that do not
match anything in the LLRequest struct.  This should guard against
unintentional typos when users are creating config files.

This commit ALSO uses the new %w verb in fmt.Errorf and therefore
makes it backward imcompatible with Go versions less than 1.13.